Project_2/src/requirements.txt
astroid==2.15.6
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
dill==0.3.6
docx==0.2.4
Flask==2.3.2
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
lxml==4.9.3
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
Pillow==10.0.0
platformdirs==3.9.1
pylint==2.17.4
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
tomlkit==0.11.8
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
wrapt==1.15.0
